reworking to conditionally render multichain stuff everywhere based on flag
