+++ src/components/WalletSelector/WalletButton.test.tsx
@@ -4,7 +4,7 @@
 // @ts-nocheck
 
 import { fireEvent, render, screen } from '@testing-library/react';
-import WalletButton from 'components/WalletSelector/DeprecatedWalletButton';
+import WalletButton from 'components/WalletSelector/WalletButton';
 import { injected } from 'connectors/index';
 
 const activate = jest.fn();
