+++ src/components/WalletSelector/multichain/MultichainWalletSelector.tsx
@@ -1,0 +1,230 @@
+import styled from 'styled-components';
+import { UnsupportedChainIdError } from '@web3-react/core';
+import { injected } from 'connectors/index';
+import { useCallback, useMemo, useState } from 'react';
+import { BaseM, TitleS } from 'components/core/Text/Text';
+import { Button } from 'components/core/Button/Button';
+import Spacer from 'components/core/Spacer/Spacer';
+import { ADD_WALLET_TO_USER, AUTH, CONNECT_WALLET_ONLY, ETHEREUM, WalletName } from 'types/Wallet';
+import breakpoints from 'components/core/breakpoints';
+import { UserRejectedRequestError } from '@web3-react/injected-connector';
+import AuthenticateWalletPending from '../AuthenticateWalletPending/AuthenticateWalletPending';
+import AddWalletPending from '../AddWalletPending/AddWalletPending';
+import Markdown from 'components/core/Markdown/Markdown';
+import { getUserFriendlyWalletName } from 'utils/wallet';
+import { graphql, useFragment } from 'react-relay';
+import { MultichainWalletSelectorFragment$key } from '__generated__/MultichainWalletSelectorFragment.graphql';
+import { isNotEarlyAccessError } from 'contexts/analytics/authUtil';
+import WalletButton from './WalletButton';
+import { useConnectEthereum } from './useConnectEthereum';
+import { MultichainAuthenticateWallet } from './MultichainAuthenticateWallet';
+import { ConnectionMode } from '../WalletSelector';
+
+type ErrorCode = string;
+type Web3Error = Error & { code: ErrorCode };
+type ErrorMessage = {
+  heading: string;
+  body: string;
+};
+
+// TODO: consider making these enums
+const ERROR_MESSAGES: Record<ErrorCode, ErrorMessage> = {
+  // Client-side provider errors: https://eips.ethereum.org/EIPS/eip-1193#provider-errors
+  UNSUPPORTED_CHAIN: {
+    heading: 'Authorization error',
+    body: 'The selected chain is unsupported. We currently only support the Ethereum network.',
+  },
+  REJECTED_SIGNATURE: {
+    heading: 'Signature required',
+    body: 'Please sign the message with your wallet to log in.',
+  },
+  USER_SIGNUP_DISABLED: {
+    heading: 'Coming Soon',
+    body: "We've detected a Gallery Member Card in your wallet! You'll be able to use it to create an account with us soon.\n\n For further updates, find us on Twitter or join our Discord.",
+  },
+  EXISTING_USER: {
+    heading: 'This address is already associated with an existing user.',
+    body: 'Sign in with that address',
+  },
+  UNKNOWN_ERROR: {
+    heading: 'There was an error connecting',
+    body: 'Please try again.',
+  },
+};
+
+function getErrorMessage(errorCode: string) {
+  return ERROR_MESSAGES[errorCode] ?? ERROR_MESSAGES.UNKNOWN_ERROR;
+}
+
+type Props = {
+  connectionMode?: ConnectionMode;
+  queryRef: MultichainWalletSelectorFragment$key;
+};
+
+export function MultichainWalletSelector({ connectionMode = AUTH, queryRef }: Props) {
+  const query = useFragment(
+    graphql`
+      fragment MultichainWalletSelectorFragment on Query {
+        ...AddWalletPendingFragment
+      }
+    `,
+    queryRef
+  );
+
+  const [isPending, setIsPending] = useState(false);
+  const [pendingWalletName, setPendingWalletName] = useState<WalletName>();
+
+  const userFriendlyWalletName = useMemo(
+    () => getUserFriendlyWalletName(pendingWalletName?.description ?? ''),
+    [pendingWalletName]
+  );
+
+  // Manually detected error not provided by web3 provider;
+  // we need to set this on state ourselves
+  const [detectedError, setDetectedError] = useState<Web3Error>();
+
+  // Default to error displayed by web3 provider and fall back
+  // to manually set error. since not all errors come with an
+  // error code, we'll add them as they come up case-by-case
+  // const displayedError = useMemo(() => {
+  //   // error is an error from the web3 provider
+  //   if (error) {
+  //     if (error instanceof UnsupportedChainIdError) {
+  //       return getErrorMessage('UNSUPPORTED_CHAIN');
+  //     }
+
+  //     if (error instanceof UserRejectedRequestError) {
+  //       return getErrorMessage('REJECTED_SIGNATURE');
+  //     }
+
+  //     return getErrorMessage('UNKNOWN_ERROR');
+  //   }
+
+  //   // detectedError is an error we manually set elsewhere in the app
+  //   if (detectedError) {
+  //     // Handle error from server
+  //     if (detectedError.code === 'GALLERY_SERVER_ERROR') {
+  //       if (isNotEarlyAccessError(detectedError.message)) {
+  //         return {
+  //           heading: account?.toLowerCase(),
+  //           body: 'Your wallet address is not on the **Early Access Allowlist**. To get onto the allowlist, visit our [FAQ](https://gallery-so.notion.site/Gallery-FAQ-b5ee57c1d7f74c6695e42c84cb6964ba#6fa1bc2983614500a206fc14fcfd61bf) or reach out to us on [Discord](discord.gg/vBqBEH8GaM).',
+  //         };
+  //       }
+
+  //       return {
+  //         heading: 'Authorization error',
+  //         body: detectedError.message,
+  //       };
+  //     }
+
+  //     return getErrorMessage(detectedError.code);
+  //   }
+
+  //   return null;
+  // }, [error, detectedError, account]);
+
+  const setToPendingState = useCallback((walletName: WalletName) => {
+    setIsPending(true);
+    setPendingWalletName(walletName);
+  }, []);
+
+  // const retryConnectWallet = useCallback(() => {
+  //   setIsPending(false);
+  //   // setDetectedError(undefined);
+  // }, []);
+
+  const connectEthereum = useConnectEthereum();
+
+  // if (displayedError) {
+  //   return (
+  //     <StyledWalletSelector>
+  //       <TitleS>{displayedError.heading}</TitleS>
+  //       <Spacer height={16} />
+  //       <StyledBody>
+  //         <Markdown text={displayedError.body} />
+  //       </StyledBody>
+  //       <StyledRetryButton onClick={retryConnectWallet}>Try Another Wallet</StyledRetryButton>
+  //     </StyledWalletSelector>
+  //   );
+  // }
+
+  if (isPending && pendingWalletName) {
+    // if (connectionMode === ADD_WALLET_TO_USER) {
+    //   return (
+    //     <StyledWalletSelector>
+    //       <AddWalletPending
+    //         queryRef={query}
+    //         setDetectedError={setDetectedError}
+    //         pendingWallet={pendingWallet}
+    //         userFriendlyWalletName={userFriendlyWalletName}
+    //         walletName={pendingWalletName}
+    //       />
+    //     </StyledWalletSelector>
+    //   );
+    // }
+
+    if (connectionMode === CONNECT_WALLET_ONLY) {
+      return (
+        <StyledWalletSelector>
+          <div>
+            <TitleS>Connect with {userFriendlyWalletName}</TitleS>
+            <Spacer height={8} />
+            <BaseM>Approve your wallet to connect to Gallery.</BaseM>
+          </div>
+        </StyledWalletSelector>
+      );
+    }
+
+    if (connectionMode === AUTH) {
+      return (
+        <StyledWalletSelector>
+          <MultichainAuthenticateWallet
+            walletName={pendingWalletName}
+            setDetectedError={setDetectedError}
+          />
+        </StyledWalletSelector>
+      );
+    }
+  }
+
+  return (
+    <StyledWalletSelector>
+      <WalletButton
+        label="Ethereum"
+        // TODO: ethereum icon
+        icon="metamask"
+        onClick={() => {
+          connectEthereum();
+          setToPendingState(ETHEREUM);
+        }}
+      />
+      {/* Just fiddling with an alternative to "more wallets coming soon" */}
+      <WalletButton disabled label="Tezos" icon="gnosis_safe" />
+      <WalletButton disabled label="Solana" icon="gnosis_safe" />
+      <WalletButton disabled label="Gnosis Safe" icon="gnosis_safe" />
+    </StyledWalletSelector>
+  );
+}
+
+const StyledWalletSelector = styled.div`
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  width: 320px;
+  justify-content: center;
+
+  @media only screen and ${breakpoints.mobileLarge} {
+    width: 400px;
+    max-width: 480px;
+  }
+`;
+
+const StyledBody = styled(BaseM)`
+  margin-bottom: 30px;
+  white-space: pre-wrap;
+`;
+
+const StyledRetryButton = styled(Button)`
+  width: 200px;
+  align-self: center;
+`;
